requirements.txt
python-telegram-bot==13.7
instagrapi
aiohttp
aiofiles
python-dotenv